+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added "mod", "rem", "round", "round-up", "round-down", and "round-to-zero" SizeFunc functions
 * Added ModSize, RemSize, RoundSize, RoundUpSize, RoundDownSize, and RoundToZeroSize functions
 * Added Start, Stop, Pause, and Resume methods to Animation interface
+* Added "transform" property and Transform interface
 * Added OpenRawResource function
 
 # v0.16.0